refactored all filter to use doubles

+++ src/main/java/fact/filter/MotionDiff.java
@@ -8,7 +8,7 @@ import stream.annotations.Description;
 import fact.Constants;
 import fact.utils.SimpleFactEventProcessor;
 /**
- * This operator calculates between data[i] and data[i+offset] for each pixel in each event and stores the result as a float array named outputKey. <\br>
+ * This operator calculates between data[i] and data[i+offset] for each pixel in each event and stores the result as a double array named outputKey. <\br>
  * if i+offset is greater or smaller the current window the first respectively the last value will be continued. 
  *  
  *@author Kai Bruegge &lt;kai.bruegge@tu-dortmund.de&gt;
@@ -16,15 +16,15 @@ import fact.utils.SimpleFactEventProcessor;
  */
 
 @Description(group = "Data Stream.FACT")
-public class MotionDiff extends SimpleFactEventProcessor<float[], float[]> {
+public class MotionDiff extends SimpleFactEventProcessor<double[], double[]> {
 	static Logger log = LoggerFactory.getLogger(MotionDiff.class);
 	
 	private int offset = 0;
 	@Override
-	public float[] processSeries(float[] data) {
-		float[] result =  data;
+	public double[] processSeries(double[] data) {
+		double[] result =  data;
 		if(! key.equals(outputKey)){
-			result = new float[data.length];
+			result = new double[data.length];
 		}
